+++ CHANGES.md
@@ -7,9 +7,12 @@ version 3.0.0 (unreleased)
     - Added `evidenceIds` to `ReportedVariant` and `ReportedStructuralVariant` (optional field)
     - Added new file types
     - Added md5Sum to `file` Record
+    - Script to migrate models to new version has been added
 
 * Major Changes:
+    - More than one version is supported, currently 2.1.0 and 3.0.0
     - Added `ClinicalReportRD` a new model to communicate the clinical reports
+    - Added `AuditLog` a new model to communicate Changes in the cases
     - Removed `VirtualPanel` model
     - `genotype` is now an enumeration
     - `ModeOfInheritance` removed from InterpretationRequest